+++ src/__tests__/ReactApp.test.tsx
@@ -9,8 +9,8 @@ import { AppContext } from "src/context/app/app.context";
 import { DEFAULT_TRACKER_DATA } from "src/constants/defaultTrackerData";
 import { TrackerData, TrackerSettings } from "src/types";
 
-jest.mock("react-i18next", () => ({
-  ...jest.requireActual("react-i18next"),
+jest.mock("src/localization/useTranslation", () => ({
+  ...jest.requireActual("src/localization/useTranslation"),
   useTranslation: jest.fn(() => ({
     t: (key: string) => key,
     i18n: {